Sowmika, Bindu | MOBN-657 | Remove unused validators and rename package
