--- omod/src/main/java/org/openmrs/module/appointments/web/validators/Validator.java
@@ -1,11 +1,0 @@
-package org.openmrs.module.appointments.web.validators;
-
-import java.util.ArrayList;
-import java.util.List;
-
-public interface Validator<T> {
-
-    boolean validate(T objectToValidate);
-
-    String getError();
-}
