+++ omod/src/test/java/org/openmrs/module/appointments/web/controller/RecurringAppointmentsControllerTest.java
@@ -18,7 +18,7 @@ import org.openmrs.module.appointments.web.contract.RecurringPattern;
 import org.openmrs.module.appointments.web.mapper.AppointmentRecurringPatternUpdateService;
 import org.openmrs.module.appointments.web.mapper.RecurringAppointmentMapper;
 import org.openmrs.module.appointments.web.service.impl.RecurringAppointmentsService;
-import org.openmrs.module.appointments.web.validators.impl.RecurringPatternValidator;
+import org.openmrs.module.appointments.web.validators.RecurringPatternValidator;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
