+++ omod/src/test/java/org/openmrs/module/appointments/web/validators/RecurringPatternValidatorTest.java
@@ -1,4 +1,4 @@
-package org.openmrs.module.appointments.web.validators.impl;
+package org.openmrs.module.appointments.web.validators;
 
 import org.junit.Before;
 import org.junit.Rule;
